Trying to reuse code. But failed.
I tried to reuse existing code(code for handling login logic) for
implementation of login logic for admin users.
But I later found that if I insisted to do so, to enable existing code
to handle new situation, almost every line of existing code is going to
introduce a parameter, which, when specified, dictates what is going to
happen under specific situation (login or login for admin).
This commit  is used to show the complexity involved if I insist to
reuse code. Next commit will present a non-reuse alternative.